// src/app/models/fse-installation-details.ts
export interface Prospect {
    saveFlag: boolean;
    updateFlag: boolean;
    deleteFlag: boolean;
    getFlag: boolean;
    fse_angaza_id: string;
    prospect_id: string;
    installation_picture: string;
    latitude: string;
    longitude: string;
    distance: string;
    accuracy: string;
    attempt: string;
    country: string;
    image_name: string;
    is_rejected: string;
    rejection_attempt: string;
}